pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["helix/"]

[tool.uv.sources]
helix-py = { workspace = true }

[tool.uv.workspace]
members = [
    ".",
]

[project]
name = "helix-py"
version = "0.2.25"
authors = [{ name = "HelixDB Team", email = "lukasnitzsche@yahoo.com"}]
maintainers = [{ name = "Lukas Nitzsche", email = "lukasnitzsche@yahoo.com" }]
description = "helix-db python lib + workflows"
readme = "README.md"
requires-python = ">=3.10"
liscense = { file = "LICENSE" }
keywords = [
    "artificial-intelligence",
    "machine-learning",
    "knowledge graphs",
    "rag",
]
dependencies = [
    "anthropic>=0.64.0",
    "chonkie>=1.1.1",
    "fastmcp>=2.10.6",
    "google-genai>=1.31.0",
    "numpy >= 2.0.2",
    "openai>=1.101.0",
    "openai-agents>=0.2.9",
    "pip>=25.1.1",
    "pyarrow >= 19.0.1",
    "requests>=2.32.4",
    "tqdm >= 4.67.1",
    "voyageai>=0.3.4",
]

[project.urls]
Homepage = "https://www.helix-db.com/"
Documentation = "https://docs.helix-db.com/"
Issues = "https://github.com/HelixDB/helix-py/issues"

[dependency-groups]
dev = [
    "jinja2>=3.1.6",
    "pyvis>=0.3.2",
    "spacy>=3.8.7",
    "build >= 1.2.2.post1",
    "chonkie>=1.0.10",
    "fastapi>=0.115.14",
    "pydantic>=2.11.7",
    "pymupdf4llm>=0.0.25",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
    "transformers>=4.52.4",
    "twine >= 6.1.0",
    "uvicorn>=0.34.3",
    "helix-py",
]

